Move rebootContainerd to testutil
