+++ integration/pull_test.go
@@ -48,7 +48,6 @@ import (
 	"github.com/containerd/containerd/platforms"
 	"github.com/opencontainers/go-digest"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
-	"github.com/rs/xid"
 )
 
 // TestSnapshotterStartup tests to run containerd + snapshotter and check plugin is
@@ -842,67 +841,6 @@ func buildSparseIndex(sh *shell.Shell, src imageInfo, minLayerSize int64) string
 		X("soci", "create", src.ref, "--min-layer-size", fmt.Sprintf("%d", minLayerSize), "--platform", platforms.Format(src.platform)).
 		O("soci", "index", "list", "-q", "--ref", src.ref, "--platform", platforms.Format(src.platform)) // this will make SOCI artifact available locally
 	return strings.Trim(string(indexDigest), "\n")
-}
-
-func rebootContainerd(t *testing.T, sh *shell.Shell, customContainerdConfig, customSnapshotterConfig string) *testutil.RemoteSnapshotMonitor {
-	var (
-		containerdRoot    = "/var/lib/containerd/"
-		containerdStatus  = "/run/containerd/"
-		snapshotterSocket = "/run/soci-snapshotter-grpc/soci-snapshotter-grpc.sock"
-		snapshotterRoot   = "/var/lib/soci-snapshotter-grpc/"
-	)
-
-	// cleanup directories
-	testutil.KillMatchingProcess(sh, "containerd")
-	testutil.KillMatchingProcess(sh, "soci-snapshotter-grpc")
-	removeDirContents(sh, containerdRoot)
-	if isDirExists(sh, containerdStatus) {
-		removeDirContents(sh, containerdStatus)
-	}
-	if isFileExists(sh, snapshotterSocket) {
-		sh.X("rm", snapshotterSocket)
-	}
-	if snDir := filepath.Join(snapshotterRoot, "/snapshotter/snapshots"); isDirExists(sh, snDir) {
-		sh.X("find", snDir, "-maxdepth", "1", "-mindepth", "1", "-type", "d",
-			"-exec", "umount", "{}/fs", ";")
-	}
-	removeDirContents(sh, snapshotterRoot)
-
-	// run containerd and snapshotter
-	var m *testutil.RemoteSnapshotMonitor
-	containerdCmds := shell.C("containerd", "--log-level", "debug")
-	if customContainerdConfig != "" {
-		containerdCmds = addConfig(t, sh, customContainerdConfig, containerdCmds...)
-	}
-	sh.Gox(containerdCmds...)
-	snapshotterCmds := shell.C("/usr/local/bin/soci-snapshotter-grpc", "--log-level", "debug",
-		"--address", snapshotterSocket)
-	if customSnapshotterConfig != "" {
-		snapshotterCmds = addConfig(t, sh, customSnapshotterConfig, snapshotterCmds...)
-	}
-	outR, errR, err := sh.R(snapshotterCmds...)
-	if err != nil {
-		t.Fatalf("failed to create pipe: %v", err)
-	}
-	m = testutil.NewRemoteSnapshotMonitor(testutil.NewTestingReporter(t), outR, errR)
-
-	// make sure containerd and soci-snapshotter-grpc are up-and-running
-	sh.Retry(100, "ctr", "snapshots", "--snapshotter", "soci",
-		"prepare", "connectiontest-dummy-"+xid.New().String(), "")
-
-	return m
-}
-
-func removeDirContents(sh *shell.Shell, dir string) {
-	sh.X("find", dir+"/.", "!", "-name", ".", "-prune", "-exec", "rm", "-rf", "{}", "+")
-}
-
-func addConfig(t *testing.T, sh *shell.Shell, conf string, cmds ...string) []string {
-	configPath := strings.TrimSpace(string(sh.O("mktemp")))
-	if err := testutil.WriteFileContents(sh, configPath, []byte(conf), 0600); err != nil {
-		t.Fatalf("failed to add config to %v: %v", configPath, err)
-	}
-	return append(cmds, "--config", configPath)
 }
 
 // TestRpullImageThenRemove pulls and rpulls an image then removes it to confirm fuse mounts are unmounted
